DEVEXP-206: Now supporting auth via a basic header
Did a fair amount of cleanup in dbClientManager to hopefully make it easier to understand and read. 

One key rename is from "req.marklogicUsername" to "req.markLogicClientCacheKey" to make it clear that the purpose of this value is for retrieving a client instance from the cache maintained by dbClientManager. 

Added two more configs - one for testing basic auth, and one for testing what happens with basic with when there's an invalid port.