Removing docker example
Example is out-of-date and no longer works. Not looking to maintain this in the immediate future, will reconsider if there's demand for an example project, but I'm hoping the docs are good enough now for Koop and GDS that a user who is savvy with Docker can easily put together their own Dockerfile based on their preferences.